change folders name. Edit uninstall.sh

+++ dpm403_2/uninstall.sh
@@ -1,4 +1,11 @@
-cd -- "$(find -iname logs -type d)"
-cd ..
-rm -rf -- "$(pwd -P)"
-cd ..
+FOLDER="$(find -iname logs -type d)"
+
+if [ -z "${FOLDER}" ]
+then
+    echo "Source is not unzipped yet."
+else
+    cd "$FOLDER"
+    cd ..
+    rm -rf -- "$(pwd -P)"
+    cd ..
+fi
